LWM's Deep Storage Support
Adds support for LWM's Deep Storage storage units, base version overrides LWM sizes making all the hold the same as the vanilla shelf. The LWM folder one, that only loads if LWM in loaded, allows LWM settings to effect base (normal quality) storage and Quality to modify that.
In game LWM menus do not reflect Quality sizes and show the max size as it is set in the LWM settings, but will actually accept more/less than that based on Quality.